chore: isolate telegram bot
