--- modules/calendar_utils.py
@@ -1,1 +1,0 @@
-../utils/calendar_utils.py
