--- modules/formatter.py
@@ -1,1 +1,0 @@
-../utils/formatter.py
