--- modules/persian_utils.py
@@ -1,1 +1,0 @@
-../utils/persian_utils.py
