+++ telegram_bot/config_local.py
@@ -1,0 +1,7 @@
+from .config import *  # noqa: F401,F403
+
+# Local SQLite database configuration
+DB_CONFIG = {
+    "database": "finance_bot_local.db",
+    "check_same_thread": False,
+}
