+++ telegram_bot/modules/persian_utils.py
@@ -1,0 +1,1 @@
+../utils/persian_utils.py
