+++ telegram_bot/test_instructions.md
