+++ telegram_bot/tests/test_ai_direct.py
