+++ telegram_bot/tests/test_basic_ai.py
