+++ telegram_bot/tests/test_full_application.py
